requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
tensorflow==2.12.0
numpy==1.23.5
Pillow==9.5.0
opencv-python==4.7.0.72
flask-session==0.5.0
reportlab==3.6.13